// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  SUPERADMIN: 'SUPERADMIN',
  ADMIN: 'ADMIN',
  EDITOR: 'EDITOR'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserRank = {
  Sestra_miloserdiya: 'Sestra_miloserdiya',
  Chinovnik: 'Chinovnik',
  Chinovnik_voennogo_vedomstva: 'Chinovnik_voennogo_vedomstva',
  Nizhniy_chin: 'Nizhniy_chin',
  Ofitser: 'Ofitser',
  Ryadovoy: 'Ryadovoy',
  Efreytor: 'Efreytor',
  Mladshiy_unter_ofitser: 'Mladshiy_unter_ofitser',
  Starshiy_unter_ofitser: 'Starshiy_unter_ofitser',
  Feldfebel: 'Feldfebel',
  Podpraporshchik: 'Podpraporshchik',
  Praporshchik: 'Praporshchik',
  Podporuchik: 'Podporuchik',
  Poruchik: 'Poruchik',
  Shtab_kapitan: 'Shtab_kapitan',
  Kapitan: 'Kapitan',
  Podpolkovnik: 'Podpolkovnik',
  Polkovnik: 'Polkovnik',
  General_mayor: 'General_mayor',
  General_leytenant: 'General_leytenant',
  General_ot_infanterii: 'General_ot_infanterii',
  Kanonir: 'Kanonir',
  Bombardir: 'Bombardir',
  Mladshiy_feyerverker: 'Mladshiy_feyerverker',
  Starshiy_feyerverker: 'Starshiy_feyerverker',
  Vakhmistr: 'Vakhmistr',
  General_ot_kavalerii: 'General_ot_kavalerii',
  Kazak: 'Kazak',
  Mladshiy_uryadnik: 'Mladshiy_uryadnik',
  Starshiy_uryadnik: 'Starshiy_uryadnik',
  Khorunzhiy: 'Khorunzhiy',
  Sotnik: 'Sotnik',
  Podesaul: 'Podesaul',
  Esaul: 'Esaul',
  Voyskovoy_starshina: 'Voyskovoy_starshina',
  Matros: 'Matros',
  Kvartirmeyster: 'Kvartirmeyster',
  Botsman: 'Botsman',
  Konduktor: 'Konduktor',
  Kapitan_2_ranga: 'Kapitan_2_ranga',
  Kapitan_1_ranga: 'Kapitan_1_ranga',
  Kontr_admiral: 'Kontr_admiral',
  Vitse_admiral: 'Vitse_admiral',
  Admiral: 'Admiral',
  Michman: 'Michman',
  Leytenant: 'Leytenant',
  Starshiy_leytenant: 'Starshiy_leytenant',
  Kollezhskiy_registrator: 'Kollezhskiy_registrator',
  Provintsialnyy_sekretar: 'Provintsialnyy_sekretar',
  Gubernskiy_sekretar: 'Gubernskiy_sekretar',
  Kollezhskiy_sekretar: 'Kollezhskiy_sekretar',
  Titulyarnyy_sovetnik: 'Titulyarnyy_sovetnik',
  Kollezhskiy_asesor: 'Kollezhskiy_asesor',
  Nadvornyy_sovetnik: 'Nadvornyy_sovetnik',
  Kollezhskiy_sovetnik: 'Kollezhskiy_sovetnik',
  Taynyy_sovetnik: 'Taynyy_sovetnik',
  Deystvitelnyy_statskiy_sovetnik: 'Deystvitelnyy_statskiy_sovetnik',
  Statskiy_sovetnik: 'Statskiy_sovetnik',
  Deystvitelnyy_taynyy_sovetnik: 'Deystvitelnyy_taynyy_sovetnik'
} as const

export type UserRank = (typeof UserRank)[keyof typeof UserRank]


export const PostStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]